solo/requirements.txt
base58==0.2.4
boto3==1.4.1
botocore==1.4.70
click==6.6
docutils==0.12
Flask==0.11.1
futures==3.0.5
hjson==2.0.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.10.0
MarkupSafe==0.23
placebo==0.8.1
python-dateutil==2.6.0
python-dotenv==0.6.0
python-slugify==1.2.1
PyYAML==3.12
requests==2.11.1
s3transfer==0.1.9
six==1.10.0
soco==0.12
tqdm==4.9.0
troposphere==1.8.2
Unidecode==0.4.19
Werkzeug==0.11.11
wsgi-request-logger==0.4.6
xmltodict==0.10.2
zappa==0.28.3
